requirements.txt
# FastAPI and async
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.2.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
aiosqlite>=0.19.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Celery and Redis
celery>=5.3.0
redis>=5.0.0
flower>=2.0.0

# Video/Audio Processing
moviepy>=1.0.3
imageio>=2.31.0
imageio-ffmpeg>=0.4.9
pydub>=0.25.1

# AI Integration
openai>=1.12.0

# Image Processing - Updated for Python 3.13 compatibility
Pillow>=10.4.0
numpy>=1.26.0

# Security & Storage
keyring>=24.3.0
cryptography>=41.0.0

# HTTP & Networking
requests>=2.31.0
httpx>=0.25.0
websockets>=12.0

# Validation and Settings - Using pydantic v1 for compatibility
pydantic>=1.10.0,<2.0.0
pydantic-settings>=1.2.0
python-dotenv>=1.0.0

# Other utilities
python-dateutil>=2.8.2
pytz>=2023.3
email-validator>=2.0.0

# Monitoring (optional)
sentry-sdk>=1.40.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.12.0
flake8>=6.0.0